backend/requirements.txt
# backend/requirements.txt

fastapi
uvicorn[standard]

langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52

ollama
chromadb
sentence-transformers
python-dotenv
pydantic-settings # 现代 FastAPI 项目管理配置的最佳实践
pypdf 
python-multipart
# ... other dependencies as needed
playwright  # <--- Add this line
python-dotenv
celery # For background tasks
redis  # The client for the Redis broker